Delete button on room reservation fixed

+++ app/Http/Livewire/Admin/RoomReservations/Index.php
@@ -18,29 +18,42 @@ class Index extends Component
 
         //$room_reservations = RoomReservation::paginate(10);
         // dd($room_reservations);
-        return view('livewire.admin.room-reservations.index', ['room_reservations'=> $room_reservations]);
+        return view('livewire.admin.room-reservations.index', ['room_reservations' => $room_reservations]);
     }
 
     public function checkoutRoom($id)
     {
         $room_reservation = RoomReservation::find($id);
-        
+
         if ($room_reservation) {
             // Delete the room reservation
             $room_reservation->delete();
-            
+
             // Delete the corresponding room selection
             $roomSelection = RoomSelection::where('room_reservation_id', $id)->first();
             if ($roomSelection) {
                 $roomSelection->delete();
             }
-            
+
             session()->flash('success', 'Room has been checked out successfully.');
         }
     }
 
-    public function deleteRoomReservation($id){
+    public function deleteRoomReservation($id)
+    {
         $reservation = RoomReservation::where('id', $id)->first();
-        $reservation->delete();
+
+        if ($reservation) {
+            // Delete the room reservation
+            $reservation->delete();
+
+            // Delete the corresponding room selection
+            $roomSelection = RoomSelection::where('room_reservation_id', $id)->first();
+            if ($roomSelection) {
+                $roomSelection->delete();
+            }
+
+            session()->flash('success', 'Room has been checked out successfully.');
+        }
     }
 }
